pyproject.toml
[project]
name = "llmops-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "doc2text>=0.2.4",
    "docx2txt==0.9",
    "faiss-cpu>=1.13.2",
    "fastapi==0.115.6",
    "ipykernel==6.30.0",
    "jinja2==3.1.4",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-google-genai==2.1.8",
    "langchain-groq==0.3.6",
    "langchain-openai==0.2.10",
    "langsmith>=0.7.7",
    "openai>=1.109.1",
    "pandas>=3.0.2",
    "pypdf>=6.9.2",
    "pytest>=9.0.3",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "sentence-transformers>=5.3.0",
    "structlog>=25.5.0",
    "uvicorn==0.32.1",
]
